opt-algorithm/requirements.txt
absl-py==2.1.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
Django==4.2.5
djangorestframework==3.14.0
executing==2.0.0
fonttools==4.43.1
geojson==2.5.0
gmaps==0.9.0
googlemaps==4.10.0
idna==3.4
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
joblib==1.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.1
ortools==9.8.3296
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
polyline==2.0.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psycopg2==2.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.2.0
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
widgetsnbextension==4.0.9
scikit-learn==1.3.2
scikit-optimize==0.10.1
geopy